+++ packages/expo-splash-screen/src/index.native.ts
@@ -1,9 +1,9 @@
-import { requireNativeModule } from 'expo-modules-core';
+import { requireOptionalNativeModule } from 'expo-modules-core';
 
-const SplashModule = requireNativeModule('ExpoSplashScreen') as {
+const SplashModule = requireOptionalNativeModule('ExpoSplashScreen') as {
   preventAutoHideAsync: () => Promise<boolean>;
   hideAsync: () => Promise<boolean>;
-};
+} | null;
 
 let _userControlledAutoHideEnabled = false;
 let _preventAutoHideAsyncInvoked = false;
@@ -17,6 +17,10 @@ let _preventAutoHideAsyncInvoked = false;
  * @private
  */
 export async function _internal_preventAutoHideAsync(): Promise<boolean> {
+  if (!SplashModule) {
+    return false;
+  }
+
   // Memoize, this should only be called once.
   if (_preventAutoHideAsyncInvoked) {
     return false;
@@ -53,6 +57,10 @@ export const _internal_maybeHideAsync = () => {
 };
 
 export function hideAsync() {
+  if (!SplashModule) {
+    return Promise.resolve(false);
+  }
+
   return SplashModule.hideAsync().catch((error: any) => {
     // Hide this very unfortunate error.
     if (